refactor: organize module by migrating GameContext to its own subdirectory
